server/requirements.txt
flask==3.0.0
gunicorn==21.2.0
google-generativeai==0.8.3
python-dotenv==1.0.0
Pillow==10.1.0